feat: add alerts
Alerts allow specific conditions or information to be highlighted at the top of
the report. This introduces a framework that allows admins to configure rules
based on the infrastructure needs. It makes use of the popular gval library and
go templates to evaluate conditions on the fly. When an alert condition is
evaluated to true, the alert is fired.

Example,

```
- message: One or more OSDs down
  when: sumUint(Status.OSDMap.OSDs, "Up") != len(Status.OSDMap.OSDs)
  severity: warning
```

Also, we now use MongoDB to store the scraped metrics, instead of generating
and storing html files in Kubernetes persistent volume. This will allows us to
make changes to UI without breaking compatibility with older rinc versions.
Also, working with a database is much nicer in general than storing files on
disk.